+++ CHANGELOG.md
@@ -6,6 +6,7 @@ adheres to [Semantic Versioning][semver].
 
 ## [Unreleased][unreleased]
 
+## [v0.0.5][v0.0.5] - 2018-06-222
 ## Changed
 - Refactored and separated match parsing and DB logic code.
 - Moved all goal and DynamoDB related code from `goal-checker` to
@@ -36,8 +37,9 @@ adheres to [Semantic Versioning][semver].
 
 [keepachangelog]: http://keepachangelog.com/en/1.0.0/
 [semver]: http://semver.org/spec/v2.0.0.html
-[unreleased]: https://github.com/pawelad/just-scored/compare/v0.0.4...HEAD
+[unreleased]: https://github.com/pawelad/just-scored/compare/v0.0.5...HEAD
 [v0.0.1]: https://github.com/pawelad/just-scored/releases/tag/v0.0.1
 [v0.0.2]: https://github.com/pawelad/just-scored/releases/tag/v0.0.2
 [v0.0.3]: https://github.com/pawelad/just-scored/releases/tag/v0.0.3
 [v0.0.4]: https://github.com/pawelad/just-scored/releases/tag/v0.0.4
+[v0.0.5]: https://github.com/pawelad/just-scored/releases/tag/v0.0.5